Run StrongMigrations checks after install
Since we are using the [Strong Migrations](https://github.com/ankane/strong_migrations) gem
and because of that if we would run the migrations automatically it will break.
We installed Solidus without running the migrations (and some of the other steps that need the migrations) and
configure Strong Migrations to start checking after the installed migrations.